Fixed build at Linux. Check map files existence at mangosd start. Now mangosd will shutdown if map directory not configured correctly. Fixed one more crash related to corpse erasing (my error in corpse changes).
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1995 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/shared/Common.h
@@ -113,7 +113,7 @@
 
 #define GUID_HIPART(x)   (*(((uint32*)&(x))+1))
 #define GUID_LOPART(x)   (*((uint32*)&(x)))
-#define MAKE_GUID(l, h)  uint64((uint32(l)) | ((uint64(uint32(h))) << 32))
+#define MAKE_GUID(l, h)  uint64( uint32(l) | ( uint64(h) << 32 ) )
 
 #define atol(a) strtoul( a, NULL, 10)
